Extract Player & State as standalone classes
