+++ core/src/com/mygdx/game/Player/State.java
@@ -1,0 +1,13 @@
+package com.mygdx.game.Player;
+
+public enum State {
+    idle, walk, jumpUp, jumpFall;
+
+    public boolean ground () {
+        return this == idle || this == walk;
+    }
+
+    public boolean air () {
+        return !ground();
+    }
+}
